Added login and signup alerts
