+++ public/js/login.js
@@ -1,24 +1,29 @@
+
 const loginFormHandler = async (event) => {
   event.preventDefault();
-
-  const email = document.querySelector('#email-login').value.trim();
-  const password = document.querySelector('#password-login').value.trim();
-
-  if (email && password) {
-    const response = await fetch('/api/users/login', {
-      method: 'POST',
-      body: JSON.stringify({ email, password }),
-      headers: { 'Content-Type': 'application/json' },
-    });
-
-    if (response.ok) {
-      document.location.replace('/');
-    } else {
-      alert('Failed to log in');
-    }
-  }
+      alert('Failed to login');
 };
+
+
+
+
+const signupFormHandler = async (event) => {
+  event.preventDefault();
+
+
+      alert('Failed to sign up.');
+  
+
+};
+
+
+
 
 document
   .querySelector('.login-form')
   .addEventListener('submit', loginFormHandler);
+
+  document
+  .querySelector('.signup-form')
+  .addEventListener('submit', signupFormHandler);
+
