+++ public/js/signup.js
@@ -1,0 +1,15 @@
+
+const signupFormHandler = async (event) => {
+    event.preventDefault();
+  
+  
+        alert('Failed to sign up.');
+    
+  
+  };
+  
+  
+  
+    document
+    .querySelector('.signup-form')
+    .addEventListener('submit', signupFormHandler);
