+++ views/login.handlebars
@@ -23,7 +23,8 @@
         <img class="mx-auto h-32 w-auto shadow-xl rounded-full" src="/images/owlsmaincircle.png" alt="Your Company">
         <h2 class="mt-10 text-center text-2xl font-bold leading-9 tracking-tight text-gray-900">Sign in to your account</h2>
       </div>
-      <form class="space-y-6" action="#" method="POST">
+
+      <form class="form login-form space-y-6">
         <div class="relative -space-y-px rounded-md shadow-sm">
           <div class="pointer-events-none absolute inset-0 z-10 rounded-md ring-1 ring-inset ring-gray-300"></div>
           <div class="shadow-xl">
@@ -58,4 +59,8 @@
       </p>
     </div>
   </div>
-  
+
+
+
+
+<script src="./js/login.js"></script>
